requirements.txt
streamlit>=1.22.0
openai>=0.27.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
plotly>=5.13.0 